refactor: move contract-producer + materialization-source types to homeboy-extension-contract
Seventh sub-cut of the extension-contract campaign (Stage 1, wiki 12840).

Extracts two self-contained pure-type clusters from manifest.rs into a new
contract extension_contract_producer module:
- ExtensionContractProducer (+ Invocation, Output, OutputKind, Phase)
- ExtensionMaterializationSourceContract (+ HelperManifestRef, SourceKind)

Plus their two serde-default helper fns and the two schema constants
(EXTENSION_CONTRACT_PRODUCER_SCHEMA, EXTENSION_MATERIALIZATION_SOURCE_SCHEMA),
which are referenced by extension tests and the cli command-contract registry.

All types are impl-free with self-closed field deps. manifest.rs re-exports
the cluster (pub use), keeping the crate::extension::* surface stable.
